Update 2.80 (backwards compatible) + extra...
Convert formatting to PEP8 standards (mostly)
Added continuous mouse drag. 

Removed mouse defaults, and opting to only remember the key pressed on start. 
This will allow any hotkey to be set for this.  
I personally set it to SPACE (as CLICK_DRAG) in frames, and switch SPACE (PRESS) to SPACE (CLICK). This allows regular playback or sliding with the same button.  
Also tweaked code to work outside the 3D View.

I'd also like there to be an easy option somewhere that adds a hotkey, rather than it being completely manual.