Alternate code for 2.80 update
This is a version of code I use. 
It has a register_props() function to register properties to a Operator/etc. 
Normally if you want to go between 2.80 and 2.7x, you have to duplicate the property assignments, because assigning props has to be like "X: Prop()" or you get a warning message, whereas before it was  always "X = Prop()". 
And you can't have the 2.80 version at all in 2.7x.

So, the register_props() function returns executable text to properly declare properties for both versions, allowing you to set the code like normal, only once. 

This is purely a preference and the code is functionally the same as the other version.